// src/app/api/weather/types/openweathermap/DailyForecastResponse.ts
import { Coord } from "~/app/api/weather/types/openweathermap/Coord";
import { Weather } from "~/app/api/weather/types/openweathermap/Weather";

export interface DailyForecastResponse {
  city: City;
  cod: string;
  message: number;
  cnt: number;
  list: Forecast[];
}

export interface City {
  id: number;
  name: string;
  coord: Coord;
  country: string;
  population: number;
  timezone: number;
}

export interface Forecast {
  dt: number;
  sunrise: number;
  sunset: number;
  temp: Temp;
  feels_like: FeelsLike;
  pressure: number;
  humidity: number;
  weather: Weather[];
  speed: number;
  deg: number;
  gust: number;
  clouds: number;
  pop: number;
  rain?: number;
  snow?: number;
}

export interface FeelsLike {
  day: number;
  night: number;
  eve: number;
  morn: number;
}

export interface Temp {
  day: number;
  min: number;
  max: number;
  night: number;
  eve: number;
  morn: number;
}